+++ crates/jsonschema/src/keywords/dependencies.rs
@@ -112,11 +112,28 @@ impl<F: Json> Validate<F> for DependenciesValidator<F> {
         tracker: Option<&RefTracker>,
         ctx: &mut ValidationContext,
     ) -> EvaluationResult {
+        self.evaluate_with_location(instance, location, &location.into(), tracker, ctx)
+    }
+
+    fn evaluate_with_location(
+        &self,
+        instance: &F::Node<'_>,
+        location: &LazyLocation,
+        instance_location: &Location,
+        tracker: Option<&RefTracker>,
+        ctx: &mut ValidationContext,
+    ) -> EvaluationResult {
         if let Some(object) = instance.as_object() {
             let mut children = Vec::new();
             for (property, dependency) in &self.dependencies {
                 if object.get(property).is_some() {
-                    children.push(dependency.evaluate_instance(instance, location, tracker, ctx));
+                    children.push(dependency.evaluate_instance_at(
+                        instance,
+                        location,
+                        instance_location,
+                        tracker,
+                        ctx,
+                    ));
                 }
             }
             EvaluationResult::from_children(children)
@@ -242,11 +259,28 @@ impl<F: Json> Validate<F> for DependentRequiredValidator<F> {
         tracker: Option<&RefTracker>,
         ctx: &mut ValidationContext,
     ) -> EvaluationResult {
+        self.evaluate_with_location(instance, location, &location.into(), tracker, ctx)
+    }
+
+    fn evaluate_with_location(
+        &self,
+        instance: &F::Node<'_>,
+        location: &LazyLocation,
+        instance_location: &Location,
+        tracker: Option<&RefTracker>,
+        ctx: &mut ValidationContext,
+    ) -> EvaluationResult {
         if let Some(object) = instance.as_object() {
             let mut children = Vec::new();
             for (property, dependency) in &self.dependencies {
                 if object.get(property).is_some() {
-                    children.push(dependency.evaluate_instance(instance, location, tracker, ctx));
+                    children.push(dependency.evaluate_instance_at(
+                        instance,
+                        location,
+                        instance_location,
+                        tracker,
+                        ctx,
+                    ));
                 }
             }
             EvaluationResult::from_children(children)
@@ -342,11 +376,28 @@ impl<F: Json> Validate<F> for DependentSchemasValidator<F> {
         tracker: Option<&RefTracker>,
         ctx: &mut ValidationContext,
     ) -> EvaluationResult {
+        self.evaluate_with_location(instance, location, &location.into(), tracker, ctx)
+    }
+
+    fn evaluate_with_location(
+        &self,
+        instance: &F::Node<'_>,
+        location: &LazyLocation,
+        instance_location: &Location,
+        tracker: Option<&RefTracker>,
+        ctx: &mut ValidationContext,
+    ) -> EvaluationResult {
         if let Some(object) = instance.as_object() {
             let mut children = Vec::new();
             for (property, dependency) in &self.dependencies {
                 if object.get(property).is_some() {
-                    children.push(dependency.evaluate_instance(instance, location, tracker, ctx));
+                    children.push(dependency.evaluate_instance_at(
+                        instance,
+                        location,
+                        instance_location,
+                        tracker,
+                        ctx,
+                    ));
                 }
             }
             EvaluationResult::from_children(children)