+++ crates/jsonschema/src/keywords/ref_.rs
@@ -63,9 +63,25 @@ impl<F: Json> Validate<F> for RefValidator<F> {
         tracker: Option<&RefTracker>,
         ctx: &mut ValidationContext,
     ) -> EvaluationResult {
+        self.evaluate_with_location(instance, location, &location.into(), tracker, ctx)
+    }
+
+    fn evaluate_with_location(
+        &self,
+        instance: &F::Node<'_>,
+        location: &LazyLocation,
+        instance_location: &Location,
+        tracker: Option<&RefTracker>,
+        ctx: &mut ValidationContext,
+    ) -> EvaluationResult {
         let child_tracker = RefTracker::new(&self.ref_suffix, &self.ref_target_base, tracker);
-        self.inner
-            .evaluate(instance, location, Some(&child_tracker), ctx)
+        self.inner.evaluate_with_location(
+            instance,
+            location,
+            instance_location,
+            Some(&child_tracker),
+            ctx,
+        )
     }
 
     /// Returns `ref_target_base` for `schema_path` output.
@@ -123,9 +139,25 @@ impl<F: Json> Validate<F> for DirectRefValidator<F> {
         tracker: Option<&RefTracker>,
         ctx: &mut ValidationContext,
     ) -> EvaluationResult {
+        self.evaluate_with_location(instance, location, &location.into(), tracker, ctx)
+    }
+
+    fn evaluate_with_location(
+        &self,
+        instance: &F::Node<'_>,
+        location: &LazyLocation,
+        instance_location: &Location,
+        tracker: Option<&RefTracker>,
+        ctx: &mut ValidationContext,
+    ) -> EvaluationResult {
         let child_tracker = RefTracker::new(&self.ref_suffix, &self.ref_target_base, tracker);
-        self.inner
-            .evaluate(instance, location, Some(&child_tracker), ctx)
+        self.inner.evaluate_with_location(
+            instance,
+            location,
+            instance_location,
+            Some(&child_tracker),
+            ctx,
+        )
     }
 
     fn canonical_location(&self) -> Option<&Location> {